Add Parcel.toParcelNode and tidy up
