+++ packages/dataparcels/src/parcel/__test__/ParcelChangeMethods-test.js
@@ -1,62 +1,6 @@
 // @flow
 import Parcel from '../Parcel';
 
-test('Parcel.data should return the Parcels data', () => {
-    var data = {
-        value: 123,
-        child: undefined
-    };
-
-    var expectedData = {
-        value: 123,
-        child: undefined,
-        key: '^',
-        meta: {}
-    };
-
-    expect(expectedData).toEqual(new Parcel(data).data);
-});
-
-test('Parcel.value should return the Parcels value', () => {
-    var data = {
-        value: 123
-    };
-    expect(new Parcel(data).value).toBe(123);
-});
-
-test('Parcel.value should return the same instance of the Parcels value', () => {
-    var myObject = {a:1,b:2};
-    var data = {
-        value: myObject
-    };
-    expect(new Parcel(data).value).toBe(myObject);
-});
-
-
-test('Parcels should be able to accept no config', () => {
-    let parcel = new Parcel();
-    expect(undefined).toEqual(parcel.value);
-    parcel.onChange(123);
-});
-
-test('Parcels should be able to accept just value in config', () => {
-    let parcel = new Parcel({
-        value: 123
-    });
-    expect(123).toEqual(parcel.value);
-    parcel.onChange(456);
-});
-
-test('Parcels should be able to accept just handleChange in config', () => {
-    let parcel = new Parcel({
-        handleChange: (parcel) => {
-            expect(456).toBe(parcel.value);
-        }
-    });
-    expect(undefined).toEqual(parcel.value);
-    parcel.onChange(456);
-});
-
 test('Parcel.setSelf() should call the Parcels handleChange function with the new parcelData', () => {
     expect.assertions(2);
 
@@ -190,46 +134,6 @@ test('Parcel.onChangeDOM() should work like onChange but take the value from eve
     });
 });
 
-test('Parcel.spread() returns an object with value and onChange', () => {
-    var data = {
-        value: 123,
-        handleChange: (parcel) => {
-            let {value} = parcel.data;
-            expect(value).toBe(456);
-        }
-    };
-
-    var parcel = new Parcel(data);
-
-    const {
-        value,
-        onChange
-    } = parcel.spread();
-
-    expect(value).toBe(parcel.value);
-    expect(onChange).toBe(parcel.onChange);
-});
-
-test('Parcel.spreadDOM() returns an object with value and onChange (onChangeDOM)', () => {
-    var data = {
-        value: 123,
-        handleChange: (parcel) => {
-            let {value} = parcel.data;
-            expect(value).toBe(456);
-        }
-    };
-
-    var parcel = new Parcel(data);
-
-    const {
-        value,
-        onChange
-    } = parcel.spreadDOM();
-
-    expect(value).toBe(parcel.value);
-    expect(onChange).toBe(parcel.onChangeDOM);
-});
-
 test('Parcel.setMeta() should call the Parcels handleChange function with the new meta merged in', () => {
     expect.assertions(3);
 
@@ -277,25 +181,6 @@ test('Parcel.setMeta() should call the Parcels handleChange function with the ne
     }).setMeta({
         abc: 123
     });
-});
-
-test('Parcel.meta should return meta', () => {
-    var meta = {
-        abc: 123,
-        def: 456
-    };
-
-    var data = {
-        value: 123,
-        handleChange: (parcel) => {
-            // the see if it is returned correctly
-            expect(meta).toEqual(parcel.meta);
-            expect(meta !== parcel.meta).toBe(true);
-        }
-    };
-
-    // first set the meta
-    var parcel = new Parcel(data).setMeta(meta);
 });
 
 test('Parcel.updateMeta() should call the Parcels handleChange function with the new meta merged in', () => {
@@ -369,28 +254,6 @@ test('Parcel.setChangeRequestMeta() should set change request meta', () => {
     });
 });
 
-test('Parcel.hasDispatched() should say if a parcel has dispatched from the current parcels path location', () => {
-    expect.assertions(6);
-
-    let p = new Parcel({
-        value: {
-            abc: 123,
-            def: 456
-        },
-        handleChange: (p2) => {
-            expect(p2.hasDispatched()).toBe(true);
-            expect(p2.get('abc').hasDispatched()).toBe(true);
-            expect(p2.get('def').hasDispatched()).toBe(false);
-        }
-    });
-
-    expect(p.hasDispatched()).toBe(false);
-    expect(p.get('abc').hasDispatched()).toBe(false);
-    expect(p.get('def').hasDispatched()).toBe(false);
-
-    p.get('abc').onChange(789);
-});
-
 test('Parcel.ping() should call the Parcels handleChange function with no change', () => {
     expect.assertions(1);
 
@@ -411,28 +274,4 @@ test('Parcel.ping() should call the Parcels handleChange function with no change
             expect(expectedData).toEqual(parcel.data);
         }
     }).ping(456);
-});
-
-test('Parcel.setInternalLocationShareData() and Parcel.getInternalLocationShareData should store data per location', () => {
-
-    let p = new Parcel({
-        value: {
-            abc: 123,
-            def: 456
-        }
-    });
-
-    expect({}).toEqual(p.getInternalLocationShareData());
-
-    p.get('abc').setInternalLocationShareData({x:1});
-    expect({x:1}).toEqual(p.get('abc').getInternalLocationShareData());
-
-    p.get('abc').setInternalLocationShareData({y:2});
-    expect({x:1, y:2}).toEqual(p.get('abc').getInternalLocationShareData());
-
-    expect({}).toEqual(p.get('def').getInternalLocationShareData());
-
-    p.get('def').setInternalLocationShareData({x:1});
-    expect({x:1}).toEqual(p.get('def').getInternalLocationShareData());
-
 });
