+++ packages/dataparcels/src/parcel/methods/ChildChangeMethods.js
@@ -1,5 +1,6 @@
 // @flow
 import type Parcel from '../Parcel';
+
 import ActionCreators from '../../change/ActionCreators';
 
 export default (_this: Parcel, dispatch: Function): Object => ({
